requirements.txt
beautifulsoup4==4.11.1
Django==3.2.7
requests==2.28.1
gunicorn==20.1.0
line-bot-sdk==1.20.0